refactor: add build:ts to the lambda template
