+++ templates/lambda/package.json
@@ -4,10 +4,12 @@
   "description": "",
   "main": "index.mjs",
   "scripts": {
-    "build:cleanup": "rm -f dist/index.js lambda.zip",
-    "build:bundle": "esbuild --bundle --outfile=./dist/index.js --platform=node --target=node22 ./index.mjs",
-    "build:zip": "zip -j lambda.zip dist/index.js",
-    "build": "pnpm build:cleanup && pnpm build:bundle && pnpm build:zip",
+    "build:cleanup": "rm -rf dist/ lambda.zip",
+    "build:ts": "tsc",
+    "build:bundle": "esbuild --bundle --outfile=./dist/bundle/index.js --platform=node --target=node22 ./dist/index.js",
+    "build:zip": "zip -j lambda.zip dist/bundle/index.js",
+    "build": "pnpm build:cleanup && pnpm build:ts && pnpm build:bundle && pnpm build:zip",
+    "check:types": "tsc --noEmit",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -15,6 +17,8 @@
   "license": "UNLICENSED",
   "type": "module",
   "devDependencies": {
+    "@tsconfig/node22": "^22.0.2",
+    "@types/node": "22.15.30",
     "esbuild": "^0.25.5",
     "typescript": "^5.8.3"
   }
