+++ templates/lambda/src/config.ts
@@ -1,0 +1,9 @@
+const {
+  NODE_ENV = 'production'
+} = process.env;
+
+export const config = {
+  env: {
+    NODE_ENV,
+  },
+};
